% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp.settings.r
\name{arg.names}
\alias{arg.names}
\title{(Internal) Supporting function which retrieves names of arguments.}
\usage{
arg.names(x, ...)
}
\arguments{
\item{x}{list of functions from which names of arguments are retrieved.}

\item{...}{other parameter names to keep.}
}
\value{
a character vector of argument names.
}
\description{
(Internal) Supporting function which retrieves names of arguments.
}
