% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partial.plot.r
\name{combine.columns}
\alias{combine.columns}
\title{(Internal) Combine columns into a character vector}
\usage{
combine.columns(data, sep)
}
\arguments{
\item{data}{a data.frame.}

\item{sep}{a character representing separator of grouping factor levels.}
}
\value{
a character vector.
}
\description{
This function combines all columns in a data.frame to make a character
vector. This function is not intended to be users.
}
