% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switch.par.r
\name{gg.colors}
\alias{gg.colors}
\title{Make ggplot like colors.}
\usage{
gg.colors(n)
}
\arguments{
\item{number}{of colors to make.}
}
\value{
a character vector containing colors.
}
\description{
Make ggplot like colors.
}
\examples{
barplot(1:30, col = gg.colors(30))
}
