% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switch.par.r
\name{make.palette}
\alias{make.palette}
\alias{make.palette.default}
\alias{make.palette.function}
\title{(Internal) Make palette of a graphic parameter.}
\usage{
make.palette(pal, x, ...)

\method{make.palette}{default}(pal, x, ...)

\method{make.palette}{`function`}(pal, x, ...)
}
\arguments{
\item{pal}{a graphic parameter making function or character vector.}

\item{x}{a vector of values that the parameter is made based on it.}

\item{...}{other parameters passed to the graphic parameter making function.}
}
\value{
If \code{x} is NULL, a non-named vector of the graphic parameter of
   length 1.
   Otherwise, a named vector of graphic parameters with unique values in
   \code{x} for the name of it.
}
\description{
(Internal) Make palette of a graphic parameter.
}
\section{Methods (by class)}{
\itemize{
\item \code{make.palette(default)}: Default S3 method, intended to be used for character vector.

\item \code{make.palette(`function`)}: S3 method for function.

}}
