% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parman.r
\docType{class}
\name{par.manager-class}
\alias{par.manager-class}
\alias{par.manager}
\title{(Internal) A reference class to handle graphic pars used for plotting.}
\description{
(Internal) A reference class to handle graphic pars used for plotting.
}
\section{Fields}{

\describe{
\item{\code{group}}{a character vector of group names made from factors specified by
x.names.}

\item{\code{col,lty,lwd,pch}}{a vector of graphic parameters or a function generate graphic
parameters. For the detail, see \code{\link{switch.par}}.}
}}

\section{Methods}{

\describe{
\item{\code{colors.for.persp(z.matrix)}}{Make color vector used for \code{\link[graphics]{persp}} funciton.
\describe{
    \item{\code{z.matrix}}{
        a matrix used for \code{z} option of \code{persp}.
    }
}}

\item{\code{initialize(group, col, lty, lwd, pch, ...)}}{Initialize class and set group field.}

\item{\code{legend.pars()}}{Prepare pars for \code{\link{pp.legend}} function.}

\item{\code{par.group(group.name = NULL)}}{Make a list of graphic parameters for each group.}

\item{\code{par.names()}}{Returns implimented par names}

\item{\code{par.obs()}}{Make a list of graphic parameters for each observation.}
}}

