% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partial.residual.r
\docType{class}
\name{partial.residual-class}
\alias{partial.residual-class}
\alias{partial.residual}
\title{(Internal) A reference class calculate partial residuals.}
\description{
(Internal) A reference class calculate partial residuals.
}
\section{Fields}{

\describe{
\item{\code{settings}}{pp.settings object having settings of the class.}

\item{\code{data}}{a numeric vector having calculated partial residual data.}
}}

\section{Methods}{

\describe{
\item{\code{calculate.residuals()}}{Calculate partial residuals.}

\item{\code{calculate.residuals.lsmeans()}}{Calculate partial residuals for lsmeans compatible models.}

\item{\code{extend.sequences(sequences)}}{Return extended 'numeric.sequences' of pp.settings.}

\item{\code{get.fit.for.large.data(x, relationship)}}{        Calculate fitted value for a observation.

        This function uses approximation using estimated relationship
        for calculating fitted value.

        \describe{
            \item{x}{a data.frame of one row representing one observation.}
            \item{relationship}{a data.frame having predicted relationship.}
        }
        }

\item{\code{get.fit.for.small.data(x)}}{        Calculate fitted value for a observation.

        \describe{
            \item{x}{a data.frame of one row representing one observation.}
        }
        }

\item{\code{initialize(settings)}}{        Initialize class and calculate residuals.

        \describe{
            \item{\code{settings}}{\code{\link{pp.settings}} object.}
        }
        }
}}

