% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp.drawer.r
\docType{class}
\name{pp.drawer-class}
\alias{pp.drawer-class}
\alias{pp.drawer}
\title{(Internal) A reference class to draw partial relationship graph.}
\description{
(Internal) A reference class to draw partial relationship graph.
}
\section{Fields}{

\describe{
\item{\code{settings}}{a \code{\link{pp.settings}} object to keep settings of the function.}
}}

\section{Methods}{

\describe{
\item{\code{draw()}}{Draw graph.}

\item{\code{draw.hist()}}{Draw histogram of explanatory variable.}

\item{\code{draw.interval()}}{Draw intervals of partial relationships.}

\item{\code{draw.interval.2d()}}{Draw intervals of partial.relationship in 2D graph.}

\item{\code{draw.relationship()}}{Draw partial relationships.}

\item{\code{draw.relationship.3d()}}{Draw 3D partial relationship.}

\item{\code{draw.residual()}}{Draw partial residuals.}

\item{\code{draw.residual.2d()}}{Draw 2D residuals.}

\item{\code{draw.residual.3d()}}{Draw 3D residuals.}

\item{\code{initialize(settings)}}{Initialize object.}

\item{\code{open.plot.window()}}{Open new plot window.}

\item{\code{xy.values()}}{Find possible x and y values.}
}}

