% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp.legend.r
\name{pp.legend}
\alias{pp.legend}
\title{Draw legend of partial plot.}
\usage{
pp.legend(settings, x, ...)
}
\arguments{
\item{settings}{\code{\link{pp.settings}} object resulted by\code{\link{partial.plot}}.}

\item{x}{position of the legend. For the detail, see
\code{\link[grahpic]{legend}} function.}

\item{...}{other graphic parameters passed to \code{\link[grahpic]{legend}}
function.}
}
\description{
Draw legend of partial plot.
}
\examples{

data(iris)
model <- lm(
    Petal.Length ~ (Sepal.Length + Petal.Width) * Species, data = iris
)
info <- partial.plot(model, c("Sepal.Length", "Species"), pch = 16)
partial.plot.legend(info, "topleft")
}
