% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp.settings.r
\docType{class}
\name{pp.settings-class}
\alias{pp.settings-class}
\alias{pp.settings}
\title{(Internal) A reference class to keep settings/information of partial.plot.}
\description{
This reference class keeps information used for partial.plot.
This class is not intended to be used by users directly.
}
\section{Fields}{

\describe{
\item{\code{adapter}}{an instance of model.adapter.}

\item{\code{model}}{a supported model object used for plotting.}

\item{\code{parman}}{an object of \code{\link{par.manager}} class.}

\item{\code{x.names}}{a character vector representing names of explanatory variables used
for plotting.}

\item{\code{x.names.factor}}{a character vector representing names of factor explanatory variables
used for plotting.}

\item{\code{x.names.numeric}}{a character vector representing names of numeric explanatory variables
used for plotting.}

\item{\code{type}}{a character literal indicating type of scale for plotting.
This is similar to type argument of many predict methods.
Possible values are "response", "link" and "prob".
If "link" is specified, partial relationship and residuals are
drawn in the scale of the linear predictor.
On the other hand, partial relationship and residuals are drawn
in the scale of the response variable if "response" is specified.
For classification models, only "prob", which calculate probability of
a specific class, can be used.}

\item{\code{positive.class}}{a class for which predicted probability is calculated.
If not specified, first class of the factor or first unique value of
the response variable is used.}

\item{\code{plot.type}}{a character literal to indicate plot type. Possible values are
"2D" and "3D".}

\item{\code{factor.levels}}{a list having levels of factors specified by x.names.}

\item{\code{group}}{a vector of grouping levels.}

\item{\code{unique.group}}{a vector of unique grouping levels.}

\item{\code{numeric.sequences}}{a list having sequences of numeric variables specified by x.names.}

\item{\code{data}}{a data.frame containing data used for plotting.}

\item{\code{fun.3d}}{the function draws 3D graphs.
Possibly \code{\link[graphics]{persp}} and
\code{\link[graphics]{images}} can work.}

\item{\code{draw.residual}}{a logical representing that partial.plot draws residual points.}

\item{\code{draw.relationship}}{a logical representing that partial.plot draws predicted relationships.}

\item{\code{draw.interval}}{a logical representing that partial.plot draws intervals of predicted
relationships.}

\item{\code{draw.hist}}{a logical specifying whether a histogram should be added to the graph.}

\item{\code{interval.levels}}{numeric vector indicating confidence level or quantiles for
predicted relationships.}

\item{\code{resolution}}{an integer specifying resolution of lines, polygons, wireframes,
and images of numeric variables.}

\item{\code{xlab}}{label of X axis.}

\item{\code{ylab}}{label of Y axis.}

\item{\code{zlab}}{label of Z axis.}

\item{\code{add}}{logical indicating whether graphic elements are added to existing plot.}

\item{\code{sep}}{a character representing separator of grouping factor levels.}

\item{\code{extrapolate}}{a logical indicating whether extrapolation is allowed for predicted
relationships.}

\item{\code{other.pars}}{a list containing other graphic parameters passed to partial.plot().}

\item{\code{n.cores}}{an integer specifing number of processes used for multiprocessing.}

\item{\code{relationship}}{a data.frame having partial relationship data.}

\item{\code{relationship.split}}{a list having partial relationship data splitted for each group.}

\item{\code{residual}}{a numeric vector having partial residual data.}

\item{\code{has.relationship}}{a logical indicating the object has partial relationship data.}

\item{\code{has.residual}}{a logical indicating the object has partial residual data.}
}}

\section{Methods}{

\describe{
\item{\code{check.data.access()}}{Check availability of data.}

\item{\code{check.intervals()}}{Check interval of predicted relationships.}

\item{\code{check.labels()}}{Test correctness of xlab and ylab.}

\item{\code{check.params()}}{Check errors in specified parameters and stop if any errors are found.}

\item{\code{check.positive.class()}}{Check whether specified positive class is in response variable.}

\item{\code{check.resolution()}}{Check resolution is integer.}

\item{\code{check.type()}}{Check specified type.}

\item{\code{check.x.names()}}{Check correctness of x.names.}

\item{\code{cluster.apply(X, FUN, ...)}}{Interface for lapply() or clusterApply()}

\item{\code{create.numeric.sequence(x.name)}}{Create sequence of an explanatory variable
based on original range of data.}

\item{\code{init.factor.levels()}}{Initialize factor levels.}

\item{\code{init.group()}}{Initialize a vector representing data splitting group.}

\item{\code{init.intervals()}}{Initialize settings of intervals.}

\item{\code{init.labels()}}{Set xlab, ylab and zlab for 2D & 3D plot.}

\item{\code{init.multiprocessing()}}{Initialize multiprocessing.}

\item{\code{init.numeric.sequences()}}{Initialize numeric sequences.}

\item{\code{init.numeric.sequences.2d()}}{Initialize numeric sequences.}

\item{\code{init.numeric.sequences.3d()}}{Initialize numeric sequences.}

\item{\code{init.plot.type()}}{Determine plot type (2D/3D).}

\item{\code{init.positive.class()}}{Initialize positive class.}

\item{\code{init.resolution()}}{Initialize resolution}

\item{\code{init.x.names()}}{Initialize names of factor and numeric explanatory variables.}

\item{\code{initialize(
  model,
  x.names,
  data = NULL,
  type = "response",
  positive.class = "",
  fun.3d = persp,
  draw.residual = TRUE,
  draw.relationship = TRUE,
  draw.interval = TRUE,
  draw.hist = FALSE,
  interval.levels = 0.95,
  resolution = NULL,
  xlab = NULL,
  ylab = NULL,
  zlab = NULL,
  add = FALSE,
  sep = " - ",
  extrapolate = FALSE,
  n.cores = NULL,
  ...
)}}{Initialize pp.settings object.
\describe{
    \item{\code{model}}{
        a supported model object used for plotting.
    }
    \item{\code{x.names}}{
        a character vector representing names of explanatory variables
        used for plotting.
    }
    \item{\code{data}}{
        a data.frame containing data used for plotting.
    }
    \item{\code{type}}{
        type of relationship to draw.
        Possible values are 'response', 'link' and 'prob'.
    }
    \item{\code{positive.class}}{
        class label for which probability is calculated.
        By default, partial.plot calculate probabilities for first
        class of the factor.
    }
    \item{\code{fun.3d}}{
        the function used for drawing 3D relationship graphs.
        Possibly, persp and image can work.
    }
    \item{\code{draw.residual}}{
        whether the partial.plot draws residual points.
    }
    \item{\code{draw.relationship}}{
        whether partial.plot draws predicted relationships.
    }
    \item{\code{draw.interval}}{
        whetehr partial.plot draws interval of predicted relationships.
    }
    \item{\code{draw.hist}}{
        whetehr partial.plot draws a histogram of explanatory variable.
    }
    \item{\code{interval.levels}}{
        numeric vector specifying level of confidence
        intervals/quantile.
    }
    \item{\code{resolution}}{
        an integer specifying resolution of lines, polygons,
        wireframes, and images of numeric variables.
    }
    \item{\code{xlab}, \code{ylab}, \code{zlab}}{
        a character specifying used for label of X/Y/Z axis.
    }
    \item{\code{add}}{
        logical specifying whether graphic elements are added to the
        existing plot.
    }
    \item{\code{sep}}{
        a character representing separator of grouping factor levels.
    }
    \item{\code{exprapolate}}{
        a logical indicating whether extrapolation is allowed for
        predicted relationships.
    }
    \item{\code{n.cores}}{
        an integer representing number of processes used for
        calculation. If NULL is specified, maximum number of logical
        processors are used. This value is ignored when the models
        compatible with lsmeans are specified.
    }
    \item{\code{...}}{
        other graphic parameters.
    }
}}

\item{\code{potential.arg.names(fun)}}{        Return a character vector having names of potential
        arguments of the function.

        \describe{
            \item{fun}{function to get potential argument names.}
        }
        }

\item{\code{set.function.args(function.args = list(), fun = .self$fun.3d)}}{        Assign function arguments kept in this class and returns it in a list.

        \describe{
            \item{function.args}{
                arguments for the function call to be modified.
            }
            \item{fun}{target function to be call.}
        }
        }

\item{\code{set.relationship(object)}}{Set partial relationship data.}

\item{\code{set.residual(object)}}{Import partial residual data.}

\item{\code{update.pars(
  fun.3d,
  draw.residual,
  draw.relationship,
  draw.interval,
  draw.hist,
  xlab,
  ylab,
  zlab,
  col,
  lty,
  lwd,
  pch,
  ...
)}}{Update settings of (mainly) graphic parameters for data reusing.
\code{fun.3d}, \code{draw.residual}, \code{draw.relationship},
\code{draw.interval}, \code{xlab}, \code{ylab},
\code{zlab}, \code{col}, \code{lty}, \code{lwd}, \code{pch}
and \code{...} can be set.}
}}

