% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp.legend.r
\name{prepare.args.for.legend}
\alias{prepare.args.for.legend}
\title{(Internal) Prepare arguments used for legend.}
\usage{
prepare.args.for.legend(settings, legend.args)
}
\arguments{
\item{settings}{an object of \code{\link{pp.settings}} object having settings of
partial.plot.}

\item{legend.args}{a list containing arguments used for \code{\link[grahpic]{legend}}
function manually specified by users.}
}
\value{
A list containing arguments used for \code{\link[grahpic]{legend}}
	function
}
\description{
This function overwrite arguments of legend() that users did not specify
manually with default value.
}
